+++ glean-core/python/glean/metrics/event.py
@@ -10,7 +10,6 @@ from typing import Dict, List, Optional
 from .. import _ffi
 from .._dispatcher import Dispatcher
 from ..testing import ErrorType
-from .. import _util
 
 
 from .lifetime import Lifetime
@@ -125,7 +124,7 @@ class EventMetricType:
         if self._disabled:
             return
 
-        timestamp = _util.time_ms()
+        timestamp = _ffi.lib.glean_get_timestamp_ms()
 
         @Dispatcher.launch
         def record():