+++ glean-core/rlb/src/private/event.rs
@@ -41,6 +41,15 @@ impl<K: traits::ExtraKeys> EventMetric<K> {
             extra_keys: PhantomData,
         }
     }
+
+    /// Record a new event with a provided timestamp.
+    ///
+    /// It's the caller's responsibility to ensure the timestamp comes from the same clock source.
+    /// Use [`glean::get_timestamp_ms`](crate::get_timestamp_ms) to get a valid timestamp.
+    pub fn record_with_time(&self, timestamp: u64, extra: HashMap<i32, String>) {
+        let metric = Arc::clone(&self.inner);
+        crate::launch_with_glean(move |glean| metric.record(glean, timestamp, Some(extra)));
+    }
 }
 
 #[inherent(pub)]
@@ -48,8 +57,7 @@ impl<K: traits::ExtraKeys> traits::Event for EventMetric<K> {
     type Extra = K;
 
     fn record<M: Into<Option<HashMap<<Self as traits::Event>::Extra, String>>>>(&self, extra: M) {
-        const NANOS_PER_MILLI: u64 = 1_000_000;
-        let now = time::precise_time_ns() / NANOS_PER_MILLI;
+        let now = crate::get_timestamp_ms();
 
         // Translate from [ExtraKey -> String] to a [Int -> String] map
         let extra = extra